Added my own uniform_int_distribution
It is my first draft but I managed to improve the speed of the Fisher-Yates algorithm by a factor of 2-3. It's still slower than the base-case algorithm of the Rust code but it is very close to it for array sizes up to 2^16. Have to find out why it get slower afterwards.